+++ src/main/java/com/rogoshum/magickcore/block/ElementWoolBlock.java
@@ -38,7 +38,7 @@ public class ElementWoolBlock extends BaseBlock{
     public int getLightValue(BlockState state, IBlockReader world, BlockPos pos) {
         TileEntity tile = world.getTileEntity(pos);
         if(tile instanceof ILightSourceEntity){
-            return ((ILightSourceEntity) tile).getSourceLight();
+            return (int) ((ILightSourceEntity) tile).getSourceLight();
         }
         return 0;
     }
