+++ src/main/java/com/rogoshum/magickcore/block/MagickContainerBlock.java
@@ -62,7 +62,7 @@ public class MagickContainerBlock extends BaseBlock {
     public int getLightValue(BlockState state, IBlockReader world, BlockPos pos) {
         TileEntity tile = world.getTileEntity(pos);
         if(tile instanceof ILightSourceEntity){
-            return ((ILightSourceEntity) tile).getSourceLight();
+            return (int) ((ILightSourceEntity) tile).getSourceLight();
         }
         return 0;
     }
